Time: 3 ms (64.61%), Space: 5.9 MB (30.89%) - LeetHub

+++ 7-reverse-integer/7-reverse-integer.cpp
@@ -7,6 +7,6 @@ public:
             x /= 10;
             if(res > INT_MAX || res < INT_MIN) return 0;
         }
-        return (int)res;
+        return res;
     }
 };
